src/requirements.txt
APScheduler==3.6.3
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
black==22.1.0
cachetools==4.2.2
certifi==2021.10.8
click==8.0.4
decorator==5.1.1
envparse==0.2.0
eradicate==2.0.0
executing==0.8.3
flake8==4.0.1
flake8-black==0.3.2
flake8-bugbear==22.1.11
flake8-eradicate==1.2.0
flake8-isort==4.1.1
flake8-print==4.0.0
iniconfig==1.1.1
ipython==8.1.1
isort==5.10.1
jedi==0.18.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
peewee==3.14.10
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.1
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.11.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
sentry-sdk==1.5.7
six==1.16.0
stack-data==0.2.0
testfixtures==6.18.5
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
wcwidth==0.2.5
